requirements.txt
# Track Friction Heatmap Dependencies - Python 3.12 Compatible
# Core Computer Vision & Image Processing
opencv-python>=4.8.0
scikit-image>=0.21.0
Pillow>=10.0.0

# Machine Learning & Data Science
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
joblib>=1.3.0

# Imbalanced Learning
imbalanced-learn>=0.11.0

# COCO Dataset Processing
pycocotools>=2.0.7

# API Framework
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6

# Visualization
matplotlib>=3.7.0
scipy>=1.10.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0.0